WIP feat/webhook-configuration: Update Harbor webhook configuration and add interval option
- The trigger from webhook is successfully, but need to run using go routine in parallel
- Make more optimize, if the trigger came from webhook don't check
  other apps that not related with that images (if possible)

Modified the Makefile to update the logging level to trace and added an --interval flag to the argocd-image-updater command for periodic execution. Updated the hack/test_real_payload.sh script to reflect a new authorization token for the Harbor webhook. In
pkg/webhook/container-registry/harbor.go, replaced deprecated ioutil.ReadAll with io.ReadAll and switched logging to a more specific log package. Added a new untracked script, hack/test_sleeper.sh, which is yet to be included in the commit.

These changes enhance webhook functionality and adjust logging behavior.